LayoutTests/storage/indexeddb/prefetch-bugfix-108071-expected.txt

Test for crbug.com/108071

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = window.indexedDB || window.webkitIndexedDB || window.mozIndexedDB;
PASS indexedDB == null is false
indexedDB.deleteDatabase(DBNAME)
indexedDB.open(DBNAME)
db = event.target.result
db.setVersion('new version')
setVersionSuccess():
trans = event.target.result
PASS trans !== null is true
objectStore = db.createObjectStore('store', {keyPath: 'id'})

resetObjectStore():
objectStore.clear()
objectStore.add({id: 0, name: "Alpha"})
objectStore.add({id: 1, name: "Bravo"})
objectStore.add({id: 2, name: "Charlie"})
objectStore.add({id: 3, name: "Delta"})
objectStore.add({id: 4, name: "Echo"})

iterateAndDeleteFirstElement():
trans = db.transaction(['store'], webkitIDBTransaction.READ_WRITE)
trans.objectStore('store').openCursor()
0: Alpha
trans.objectStore('store').delete(0)
1: Bravo
2: Charlie
3: Delta
4: Echo

resetObjectStore():
objectStore.clear()
objectStore.add({id: 0, name: "Alpha"})
objectStore.add({id: 1, name: "Bravo"})
objectStore.add({id: 2, name: "Charlie"})
objectStore.add({id: 3, name: "Delta"})
objectStore.add({id: 4, name: "Echo"})

prefetchAndAbort():
trans.objectStore('store').openCursor()
0: Alpha
1: Bravo
2: Charlie
3: Delta
trans.abort()
PASS Transaction aborted as expected
PASS successfullyParsed is true

TEST COMPLETE
